Create a `Converter` class and move Recap converters to it
I'm starting to clean up the `recap.schema` code a bit. Step 1 is creating a
`Converter` class that all conveters extend.

I've also moved Recap's `Parser` to be a `RecapConverter` that behaves just like
all the other converters. This has cleaned up `types.py` quite a bit.
