+++ recap/schema/models.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from pydantic import BaseModel, Extra, root_validator
 
-from recap.schema.types import Parser
+from recap.schema.converters.recap import RecapConverter
 
 
 class Type(BaseModel, extra=Extra.allow):
@@ -13,5 +13,5 @@ class Type(BaseModel, extra=Extra.allow):
     @root_validator
     def validate_recap_obj(cls, obj):
         # Parser throws exception if schema object is unparsable
-        Parser().parse_obj(obj)
+        RecapConverter().to_recap_type(obj)
         return obj
